rename scan to 'pdbscan'
