+++ pg_search/src/postgres/customscan/pdbscan/qual_inspect.rs
